Working on creating template resource on ui
