+++ lib/auth.ts
@@ -1,6 +1,7 @@
 import { AuthOptions, Session, getServerSession } from "next-auth";
 import GoogleProvider from 'next-auth/providers/google';
 import { redirect } from "next/navigation";
+import { SessionError } from "./errors/SessionError";
 
 
 
@@ -42,3 +43,13 @@ export const useServerSession = () => {
 //     router.push("/")
 //   }
 // }
+
+export const getSessionEmail = async () => {
+  const session = await useServerSession()
+  const email = session?.user?.email
+  if (!email) {
+      const err = new SessionError("Email not found in Session (Auth)")
+      return [err, null] as const
+  }
+  return [null, email] as const
+}
