+++ lib/errors/ValidateError.ts
@@ -1,0 +1,22 @@
+import { redirect } from "next/navigation"
+import { SessionError } from "./SessionError"
+import { PrismaResourceNotFound } from "./PrismaErrors"
+import { PrismaClientKnownRequestError } from "@prisma/client/runtime/library"
+import { UnknownError } from "./UnknownError"
+
+export class Validator {
+    static validateErrorOrRedirect(error: unknown) {
+        console.error(error)
+        if (error instanceof SessionError) {
+            redirect("/")
+        }
+        if (error instanceof PrismaResourceNotFound) {
+            return [error.message, null] as const
+        }
+        if (error instanceof PrismaClientKnownRequestError) {
+            return [error.message, null] as const
+        }
+        const unknownError = new UnknownError()
+        return [unknownError.message, null] as const
+    }
+}
